# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Amar Pathagar API
  description: |
    Trust-Based Community Library System API
    
    A comprehensive API for managing a community-driven book sharing platform where trust and reputation matter.
  version: 1.0.0
  contact:
    name: NesoHQ
    url: https://github.com/nesohq
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.amarpathagar.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
          enum: [admin, member]
        avatar_url:
          type: string
        bio:
          type: string
        location_lat:
          type: number
          format: double
        location_lng:
          type: number
          format: double
        location_address:
          type: string
        success_score:
          type: integer
        books_shared:
          type: integer
        books_received:
          type: integer
        total_upvotes:
          type: integer
        total_downvotes:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UserReview:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reviewer_id:
          type: string
          format: uuid
        reviewer:
          $ref: '#/components/schemas/User'
        reviewee_id:
          type: string
          format: uuid
        reviewee:
          $ref: '#/components/schemas/User'
        book_id:
          type: string
          format: uuid
        behavior_rating:
          type: integer
          minimum: 1
          maximum: 5
        book_condition_rating:
          type: integer
          minimum: 1
          maximum: 5
        communication_rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
paths:
...
  /reviews:
    post:
      summary: Create user review
      description: Write a review for another user based on book exchange experience
      tags:
        - Reviews
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reviewee_id
              properties:
                reviewee_id:
                  type: string
                  format: uuid
                  description: ID of the user being reviewed
                book_id:
                  type: string
                  format: uuid
                  description: ID of the book related to this review
                behavior_rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  description: Rating for user behavior
                book_condition_rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  description: Rating for book condition
                communication_rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  description: Rating for communication
                comment:
                  type: string
                  description: Optional review comment
      responses:
        '201':
          description: Review created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserReview'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{id}/reviews:
    get:
      summary: Get user reviews
      description: Get all reviews received by a user
      tags:
        - Reviews
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of reviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserReview'
  /books/{id}/reviews:
    get:
      summary: Get book reviews
      description: Get all reviews related to a specific book
      tags:
        - Reviews
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of reviews for the book
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserReview'
...